refactor: changed return int to size_t

+++ ft_strlen.c
@@ -6,15 +6,15 @@
 /*   By: eandre-f <eandre-f@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/04/03 00:19:08 by eandre-f          #+#    #+#             */
-/*   Updated: 2022/04/03 00:22:37 by eandre-f         ###   ########.fr       */
+/*   Updated: 2022/04/03 02:40:50 by eandre-f         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_strlen(char *s)
+size_t	ft_strlen(char *s)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (s[i])
